Task 3
I ran the unit test. I showed errors on the getArea method. I did some
looking into the values of the points during rectangle declarations and
figured out that the point values were repeating the y values during
the declaration. This caused me to look at the point class, and correct
the declaration.

At first to spot the bug, I had to place print statements within the
constructor of rectangle, because I could not directly get access to
the point instance.  My code addition was to add a getter for the
point, so that the coder can see the actual values involved within the
calculation. This, could help you determine what class has the error
code.